+++ engine/src/main/kotlin/com/kakao/actionbase/engine/query/ActionbaseQuery.kt
@@ -1,4 +1,4 @@
-package com.kakao.actionbase.v2.engine.query
+package com.kakao.actionbase.engine.query
 
 import com.kakao.actionbase.v2.core.metadata.Direction
 import com.kakao.actionbase.v2.core.types.DataType
