+++ engine/src/test/kotlin/com/kakao/actionbase/v2/engine/v3/query/ActionbaseQueryParserSpec.kt
@@ -1,8 +1,8 @@
 package com.kakao.actionbase.v2.engine.v3.query
 
 import com.kakao.actionbase.v2.core.metadata.Direction
-import com.kakao.actionbase.v2.engine.query.ActionbaseQuery
-import com.kakao.actionbase.v2.engine.query.ActionbaseQuery.Companion.toJson
+import com.kakao.actionbase.engine.query.ActionbaseQuery
+import com.kakao.actionbase.engine.query.ActionbaseQuery.Companion.toJson
 import com.kakao.actionbase.v2.engine.sql.WherePredicate
 
 import io.kotest.core.spec.style.StringSpec
