+++ server/src/main/kotlin/com/kakao/actionbase/server/api/graph/v3/NamedQueryResult.kt
@@ -1,0 +1,7 @@
+package com.kakao.actionbase.server.api.graph.v3
+
+import com.kakao.actionbase.v2.engine.sql.QueryResult
+
+data class NamedQueryResult(
+    val items: List<QueryResult.NamedJsonFormat>,
+)
