+++ server/src/main/kotlin/com/kakao/actionbase/server/api/graph/v3/QueryController.kt
@@ -1,8 +1,7 @@
 package com.kakao.actionbase.server.api.graph.v3
 
 import com.kakao.actionbase.server.util.mapToResponseEntity
-import com.kakao.actionbase.v2.engine.query.ActionbaseQuery
-import com.kakao.actionbase.v2.engine.sql.QueryResult
+import com.kakao.actionbase.engine.query.ActionbaseQuery
 import com.kakao.actionbase.v2.engine.sql.toNamedJsonFormat
 import com.kakao.actionbase.v2.engine.v3.V3QueryService
 
@@ -28,7 +27,3 @@ class QueryController(
                 NamedQueryResult(items)
             }.mapToResponseEntity()
 }
-
-data class NamedQueryResult(
-    val items: List<QueryResult.NamedJsonFormat>,
-)
